Remove a few DbCache references (#2732)
