+++ api/src/org/labkey/api/ApiModule.java
@@ -226,11 +226,11 @@ public class ApiModule extends CodeOnlyModule
             ApiKeyManager.TestCase.class,
             AtomicDatabaseInteger.TestCase.class,
             BlockingCache.BlockingCacheTest.class,
+            CachingTestCase.class,
             ContainerDisplayColumn.TestCase.class,
             ContainerFilter.TestCase.class,
             ContainerManager.TestCase.class,
             CompareType.TestCase.class,
-            DbSchema.CachingTestCase.class,
             DbSchema.DDLMethodsTestCase.class,
             DbSchema.SchemaCasingTestCase.class,
             DbSchema.TableSelectTestCase.class,
