+++ api/src/org/labkey/api/data/TempTableInClauseGenerator.java
@@ -41,7 +41,7 @@ import java.util.TreeSet;
 public class TempTableInClauseGenerator implements InClauseGenerator
 {
     private static final Cache<String, TempTableInfo> _tempTableCache =
-            CacheManager.getStringKeyCache(100, CacheManager.MINUTE * 5, "InClauseTempTableCache");
+            CacheManager.getStringKeyCache(200, CacheManager.MINUTE * 5, "InClauseTempTableCache");
 
     /**
      * @param sql    fragment to append to
