+++ experiment/src/org/labkey/experiment/api/ExperimentServiceImpl.java
@@ -994,8 +994,6 @@ public class ExperimentServiceImpl implements ExperimentService
         Cache<String, ExpProtocolImpl> c = getProtocolCache();
         c.remove(getCacheKey(p.getLSID()));
         c.remove("ROWID/" + p.getRowId());
-        //TODO I don't think we're using a DbCache for protocols...
-        DbCache.remove(getTinfoProtocol(), getCacheKey(p.getLSID()));
     }
